refactor: extract by_continent, by_country, by_city into dedicated controllers
Break up the monolithic EventsController by extracting geography-based
event browsing into three single-responsibility controllers. Shared scoping
and pagination logic is consolidated in the EventBrowsing concern.

- Events::ByContinentController handles /:continent (HTML + RSS)
- Events::ByCountryController handles /:continent/:country_name (HTML + RSS)
- Events::ByCityController handles /:continent/:country_name/:city_name (HTML)
- EventsController reduced from 569 to 348 lines
- Tests reorganized per TEST_ARCHITECTURE.md (per-action files, correct naming)
- AGENTS.md updated with namespaced controller test guidelines
- Code review skill enhanced with test architecture compliance checklist
