+++ app/views/events/by_city/show.html.erb
