+++ app/views/events/by_continent/show.html.erb
