+++ app/views/events/by_continent/show.xml.builder
