+++ app/views/events/by_country/show.html.erb
