+++ app/views/events/by_country/show.xml.builder
